Moving 3D drawing to RegDrawer
