+++ pulser/register/_reg_drawer.py
@@ -15,6 +15,7 @@
 from __future__ import annotations
 
 from typing import Optional, Union
+from itertools import combinations
 
 import matplotlib.pyplot as plt
 import numpy as np
@@ -161,6 +162,114 @@ class RegDrawer:
             ax.axhline(0, c="grey", alpha=0.5, linestyle=":")
 
     @staticmethod
+    def _draw_3D(
+        pos: np.ndarray,
+        ids: list,
+        projection: bool = False,
+        with_labels: bool = True,
+        blockade_radius: Optional[float] = None,
+        draw_graph: bool = True,
+        draw_half_radius: bool = False,
+        are_traps: bool = False,
+    ):
+        if draw_graph and blockade_radius is not None:
+            epsilon = 1e-9  # Accounts for rounding errors
+            edges = KDTree(pos).query_pairs(blockade_radius * (1 + epsilon))
+
+        if projection:
+            labels = "xyz"
+            fig, axes = RegDrawer._initialize_fig_axes_projection(
+                pos,
+                blockade_radius=blockade_radius,
+                draw_half_radius=draw_half_radius,
+            )
+            fig.tight_layout(w_pad=6.5)
+
+            for ax, (ix, iy) in zip(axes, combinations(np.arange(3), 2)):
+                RegDrawer._draw_2D(
+                    ax,
+                    pos,
+                    ids,
+                    plane=(
+                        ix,
+                        iy,
+                    ),
+                    with_labels=with_labels,
+                    blockade_radius=blockade_radius,
+                    draw_graph=draw_graph,
+                    draw_half_radius=draw_half_radius,
+                )
+                ax.set_title(
+                    "Projection onto\n the "
+                    + labels[ix]
+                    + labels[iy]
+                    + "-plane"
+                )
+
+        else:
+            fig = plt.figure(figsize=2 * plt.figaspect(0.5))
+
+            if draw_graph and blockade_radius is not None:
+                bonds = {}
+                for i, j in edges:
+                    xi, yi, zi = pos[i]
+                    xj, yj, zj = pos[j]
+                    bonds[(i, j)] = [[xi, xj], [yi, yj], [zi, zj]]
+
+            for i in range(1, 3):
+                ax = fig.add_subplot(
+                    1, 2, i, projection="3d", azim=-60 * (-1) ** i, elev=15
+                )
+
+                ax.scatter(
+                    pos[:, 0],
+                    pos[:, 1],
+                    pos[:, 2],
+                    s=30,
+                    alpha=0.7,
+                    c="darkgreen",
+                )
+
+                if with_labels:
+                    for q, coords in zip(ids, pos):
+                        ax.text(
+                            coords[0],
+                            coords[1],
+                            coords[2],
+                            q,
+                            fontsize=12,
+                            ha="left",
+                            va="bottom",
+                        )
+
+                if draw_half_radius and blockade_radius is not None:
+                    mesh_num = 20 if len(ids) > 10 else 40
+                    for r in pos:
+                        x0, y0, z0 = r
+                        radius = blockade_radius / 2
+
+                        # Strange behavior pf mypy using "imaginary slice step"
+                        # u, v = np.pi * np.mgrid[0:2:50j, 0:1:50j]
+
+                        v, u = np.meshgrid(
+                            np.arccos(np.linspace(-1, 1, num=mesh_num)),
+                            np.linspace(0, 2 * np.pi, num=mesh_num),
+                        )
+                        x = radius * np.cos(u) * np.sin(v) + x0
+                        y = radius * np.sin(u) * np.sin(v) + y0
+                        z = radius * np.cos(v) + z0
+                        # alpha controls opacity
+                        ax.plot_surface(x, y, z, color="darkgreen", alpha=0.1)
+
+                if draw_graph and blockade_radius is not None:
+                    for x, y, z in bonds.values():
+                        ax.plot(x, y, z, linewidth=1.5, color="grey")
+
+                ax.set_xlabel("x (µm)")
+                ax.set_ylabel("y (µm)")
+                ax.set_zlabel("z (µm)")
+
+    @staticmethod
     def _register_dims(
         pos: np.ndarray,
         blockade_radius: Optional[float] = None,
@@ -193,6 +302,48 @@ class RegDrawer:
             big_side / 4, 10
         )  # Figsize is, at most, (10,10)
         fig, axes = plt.subplots(figsize=Ls)
+
+        return (fig, axes)
+
+    @staticmethod
+    def _initialize_fig_axes_projection(
+        pos: np.ndarray,
+        blockade_radius: Optional[float] = None,
+        draw_half_radius: bool = False,
+    ) -> tuple[plt.figure.Figure, plt.axes.Axes]:
+        """Creates the Figure and Axes for drawing the register projections."""
+        diffs = RegDrawer._register_dims(
+            pos,
+            blockade_radius=blockade_radius,
+            draw_half_radius=draw_half_radius,
+        )
+
+        proportions = []
+        for (ix, iy) in combinations(np.arange(3), 2):
+            big_side = max(diffs[[ix, iy]])
+            Ls = diffs[[ix, iy]] / big_side
+            Ls *= max(
+                min(big_side / 4, 10), 4
+            )  # Figsize is, at most, (10,10), and, at least (4,*) or (*,4)
+            proportions.append(Ls)
+
+        fig_height = np.max([Ls[1] for Ls in proportions])
+
+        max_width = 0
+        for i, (width, height) in enumerate(proportions):
+            proportions[i] = (width * fig_height / height, fig_height)
+            max_width = max(max_width, proportions[i][0])
+        widths = [max(Ls[0], max_width / 5) for Ls in proportions]
+        fig_width = min(np.sum(widths), fig_height * 4)
+
+        rescaling = 20 / max(max(fig_width, fig_height), 20)
+        figsize = (rescaling * fig_width, rescaling * fig_height)
+
+        fig, axes = plt.subplots(
+            ncols=3,
+            figsize=figsize,
+            gridspec_kw=dict(width_ratios=widths),
+        )
 
         return (fig, axes)
